pyproject.toml
[project]
name = "fastapi-pj"
version = "0.1.0"
description = ""
authors = [
    {name = "周仁明",email = "stevenchouis@hotmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy[asyncio] (>=2.0.48,<3.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "update (>=0.0.1,<0.0.2)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "exponent-server-sdk (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
